requirements.txt
streamlit==1.35.0
numpy==1.24.3
pillow==10.3.0
joblib==1.4.2
tensorflow-cpu==2.13.0
scikit-learn==1.0.2